HACK: sysctl: disallow override of writeout policy
"Android writeout policy was never touched since 2009.
In the meantime, most of Android devices are equipped
with over 4GB DRAM and very fast flash storages like
UFS, which becomes more like desktop or servers in 2009.
So, it'd be worth to go back to use the default kernel
configs."
  - https://android-review.googlesource.com/c/platform/system/core/+/938362

Proper way is to modify init.rc but we want to avoid
touching system.

Signed-off-by: Jesse Chan <jc@linux.com>
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>